requirements.txt
discord.py
Django
gitpython
matplotlib
texttable
pandas
requests
dblpy